+++ packages/frontend/mobile/src/components/ExploreContainer.vue
@@ -1,6 +1,7 @@
 <template>
   <div id="container">
     <strong>{{ name }}</strong>
+    <pre>anna molodec</pre>
     <p>Explore <a target="_blank" rel="noopener noreferrer" href="https://ionicframework.com/docs/components">UI Components</a></p>
   </div>
 </template>