+++ packages/frontend/rules/CHANGELOG.md
@@ -2,7 +2,6 @@
 
 ## 1.0.0 (2025-04-28)
 
-
 ### Features
 
-* **@catch-coin/shared:** Installed shared dev dependencies for frontend apps ([f0626b1](https://github.com/AvdienkoSergey/catch-coin/commit/f0626b1d1aec86a757bc2a6b9a119f83c2c05e56))
+- **@catch-coin/shared:** Installed shared dev dependencies for frontend apps ([f0626b1](https://github.com/AvdienkoSergey/catch-coin/commit/f0626b1d1aec86a757bc2a6b9a119f83c2c05e56))